Refactoring of package into submodules
