+++ model/Makefile
@@ -1,0 +1,7 @@
+install:
+	poetry install
+
+test:
+	poetry run pytest 
+
+.PHONY: install test
