+++ model/src/pyrenew/models/__init__.py
@@ -1,0 +1,1 @@
+from pyrenew.models.basicrenewal import BasicRenewalModel # noqa: F401
