+++ model/src/pyrenew/observations/__init__.py
@@ -1,0 +1,7 @@
+from pyrenew.observations.poisson import PoissonObservation #noqa: F401
+from pyrenew.observations.negativebinomial import NegativeBinomialObservation #noqa: F401
+from pyrenew.observations.infection import (
+    sample_infections_rt,
+    logistic_susceptibility_adjustment,
+    sample_infections_with_feedback
+)
