+++ model/src/pyrenew/observations/infection.py
