+++ model/src/pyrenew/observations/negativebinomial.py
@@ -5,50 +5,13 @@
 Observation helper classes
 """
 
-from abc import ABCMeta, abstractmethod
+from pyrenew.metaclasses import RandomProcess
 
 import numpyro
 import numpyro.distributions as dist
 from numpy.typing import ArrayLike
 
-
-class Observation(metaclass=ABCMeta):
-    """
-    Abstract base class for an observation
-    with a single predicted value and optional
-    auxiliary parameters governing properties
-    such as observation noise
-    """
-
-    def __init__(self):
-        """
-        Default constructor
-        """
-        pass
-
-    @abstractmethod
-    def sample(
-        self, parameter_name, predicted_value: ArrayLike, data=None, obs=None
-    ):
-        """
-        Sampling method that concrete
-        versions should implement.
-        """
-        pass
-
-
-class PoissonObservation(Observation):
-    """
-    Poisson observation process
-    """
-
-    def sample(self, parameter_name, predicted_value, data=None, obs=None):
-        return numpyro.sample(
-            parameter_name, dist.Poisson(rate=predicted_value), obs=obs
-        )
-
-
-class NegativeBinomialObservation(Observation):
+class NegativeBinomialObservation(RandomProcess):
     def __init__(
         self,
         concentration_prior: dist.Distribution,
@@ -87,3 +50,6 @@ class NegativeBinomialObservation(Observation):
             ),
             obs=obs,
         )
+
+    def validate(self):
+        return None
