+++ model/src/pyrenew/processes/__init__.py
@@ -1,0 +1,5 @@
+from pyrenew.processes.simplerandomwalk import SimpleRandomWalkProcess # noqa: F401
+from pyrenew.processes.ar import ARProcess # noqa: F401
+from pyrenew.processes.firstdifferencear import FirstDifferenceARProcess #noqa: F401
+
+
