+++ model/src/test/test_ar_process.py
@@ -1,6 +1,6 @@
 import jax.numpy as jnp
 import numpyro
-from pyrenew.process import ARProcess
+from pyrenew.processes import ARProcess
 
 
 def test_ar_can_be_sampled():
