+++ model/src/test/test_random_walk.py
@@ -1,7 +1,7 @@
 import jax.numpy as jnp
 import numpyro
 import numpyro.distributions as dist
-from pyrenew.process import SimpleRandomWalk
+from pyrenew.processes import SimpleRandomWalkProcess
 
 
 def test_rw_can_be_sampled():
@@ -9,7 +9,7 @@ def test_rw_can_be_sampled():
     Check that a simple random walk
     can be initialized and sampled from
     """
-    rw_normal = SimpleRandomWalk(dist.Normal(0, 1))
+    rw_normal = SimpleRandomWalkProcess(dist.Normal(0, 1))
 
     with numpyro.handlers.seed(rng_seed=62):
         # can sample with and without inits
@@ -27,7 +27,7 @@ def test_rw_samples_correctly_distributed():
     for step_mean, step_sd in zip(
         [0, 2.253, -3.2521, 1052, 1e-6], [1, 0.025, 3, 1, 0.02]
     ):
-        rw_normal = SimpleRandomWalk(dist.Normal(step_mean, step_sd))
+        rw_normal = SimpleRandomWalkProcess(dist.Normal(step_mean, step_sd))
 
         with numpyro.handlers.seed(rng_seed=62):
             samples = rw_normal.sample(n_samples, init=jnp.array([50.0]))
